src/lambda_outbox/requirements.txt
azure-storage-blob>=12.25.1
azure-storage-queue>=12.12.0
azure-functions>=1.22.1
azure-identity>=1.21.0
python-dotenv>=1.1.0
